seprate tests from 1 file to 2 files
